+++ CHANGES.rst
@@ -14,7 +14,7 @@ Development
 - Add more wakeup reasons
 - Add missing configuration section for HX711 to settings example.
   Thanks, @ClemensGruber.
-
+- Add basic device-interval sensors ``SystemTemperature`` and ``SystemBatteryLevel``
 
 2019-06-07 0.3.0
 ================